backend/requirements.txt
abnf==2.4.1
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiomysql==0.2.0
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncio==3.4.3
asyncpg==0.30.0
attrs==25.1.0
bcrypt==3.2.2
beautifulsoup4==4.13.3
bitarray==3.8.0
blinker==1.9.0
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.1.5
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
cytoolz==1.1.0
databases==0.9.0
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.0
email_validator==2.2.0
et_xmlfile==2.0.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
Flask==3.1.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
hexbytes==1.3.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humps==0.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.12.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
mysql==0.0.3
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.2.4
openai==2.8.1
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.3
parsimonious==0.10.0
passlib==1.7.4
propcache==0.3.0
psycopg2==2.9.11
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.1
pyunormalize==17.0.0
pywin32==311
PyYAML==6.0.2
regex==2025.11.3
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rlp==4.1.0
rsa==4.9
selenium==4.29.0
shellingham==1.5.4
siwe==4.4.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
SQLAlchemy-Utils==0.41.2
starlette==0.45.3
toolz==1.1.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
twilio==9.4.6
typer==0.15.1
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
web3==7.14.0
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
